+++ src/pint/models/timing_model.py
@@ -127,11 +127,9 @@ class MissingTOAs(ValueError):
         if isinstance(parameter_names, str):
             parameter_names = [parameter_names]
         if len(parameter_names) == 1:
-            msg = f"Parameter {parameter_names[0]} does not correspond to any TOAs"
+            msg = f"Parameter {parameter_names[0]} does not correspond to any TOAs: you might need to run `model.find_empty_masks(toas, freeze=True)`"
         elif len(parameter_names) > 1:
-            msg = (
-                f"Parameters {' '.join(parameter_names)} do not correspond to any TOAs"
-            )
+            msg = f"Parameters {' '.join(parameter_names)} do not correspond to any TOAs: you might need to run `model.find_empty_masks(toas, freeze=True)`"
         else:
             raise ValueError("Incorrect attempt to construct MissingTOAs")
         super().__init__(msg)
@@ -2522,7 +2520,7 @@ class TimingModel:
         bad_parameters = []
         for maskpar in self.get_params_of_type_top("maskParameter"):
             par = getattr(self, maskpar)
-            if "TNEQ" in str(par.name) or par.frozen:
+            if par.frozen:
                 continue
             if len(par.select_toa_mask(toas)) == 0:
                 bad_parameters.append(f"'{maskpar}, {par.key}, {par.key_value}'")
@@ -2533,6 +2531,32 @@ class TimingModel:
                 bad_parameters += e.parameter_names
         if bad_parameters:
             raise MissingTOAs(bad_parameters)
+
+    def find_empty_masks(self, toas, freeze=False):
+        """Find unfrozen mask parameters with no TOAs before trying to fit
+
+        Parameters
+        ----------
+        toas : pint.toa.TOAs
+        freeze : bool, optional
+            Should the parameters with on TOAs be frozen
+
+        Returns
+        -------
+        list
+            Parameters with no TOAs
+        """
+        bad_parameters = []
+        for maskpar in self.get_params_of_type_top("maskParameter"):
+            par = getattr(self, maskpar)
+            if par.frozen:
+                continue
+            if len(par.select_toa_mask(toas)) == 0:
+                bad_parameters.append(maskpar)
+                if freeze:
+                    log.info(f"'{maskpar}' has no TOAs so freezing")
+                    getattr(self, maskpar).frozen = True
+        return bad_parameters
 
     def setup(self):
         """Run setup methods on all components."""